// src/redux/actions/actionTypes/types.ts
export enum ActionTypes {
  FetchShelters = 'FETCH_SHELTERS',
  GetShelters = 'GET_SHELTERS',
  SetShelters = 'SET_SHELTERS',
  SetShelter = 'SET_SHELTER',
  FetchError = 'SHELTERS_LOAD_FAILED',
  SetFirstName = 'SET_FIRST_NAME',
  SetLastName = 'SET_LAST_NAME',
  SetEmail = 'SET_EMAIL',
  SetPhoneNumber = 'SET_PHONE_NUMBER',
  SetValue = 'SET_VALUE',
  SetShelterId = 'SET_SHELTER_ID',
  ErrorData = 'ERROR_DATA',
  NextPage = 'NEXT_PAGE',
  PreviousPage = 'PREVIOUS_PAGE',
  SetHelpMethod = 'SET_HELP_METHOD',
  SetFormState = 'SET_FORM_STATE',
  SubmitForm = 'SUBMIT_FORM',
}